Simplify configs a smidge more
