+++ storybook/src/stories.a11y.ts
@@ -25,6 +25,18 @@
  */
 
 import initStoryshots from '@storybook/addon-storyshots'
-import { a11y } from './storyshotsConfig'
+import { axeTest } from '@storybook/addon-storyshots-puppeteer'
+import { StoryshotsOptions } from '@storybook/addon-storyshots/dist/ts3.9/api/StoryshotsOptions'
+
+const a11y = () => {
+  const storybookUrl = 'storybook/storybook-static'
+
+  return {
+    configPath: `storybook/.storybook`,
+    framework: 'react',
+    suite: 'a11y suite',
+    test: axeTest({ storybookUrl }),
+  } as StoryshotsOptions
+}
 
 initStoryshots(a11y())
